Tracing connection events
Modify 'ConnectionId' type and 'WithConnectionId':  in some contexts we
don't knwon the remote peer address, e.g. when accept loop throws an
exception.  For that reason `ConnectionId'` type is provided which
distinguishes local and remote address types.  'MaybeAddress' and
explicit (and strict) maybe type is provided for situations where we
don't know the address, togheter with 'knowRemoteAddress` and
'withKnownRemoteAddress` functions which are useful for contravariant
tracing.

The ouroboros-network-framework and ouroboros-network packages are
updated at once.