Add some randomness to the reconnection delay
In case of a TCP sim open the handshake will fail, and both
nodes will retry at exactly the same time causing the same
failure to repeat forever. The random delay breaks the cycle.

The random delay is between -1s to +3s which is more than
then the expected RTT around the world.