Fix races related to updating PeerState
A peer can be set to cold at anytime asynchronously, this means that the
governor may ask us to promote or demote a peer that has just been set
to cold (for example the peer reset the connection).

This change makes sure that each time we're about to update to peer
status to something other than PeerCold, that we first check to see if
the peer is coold. If that is the case we fail with an exception.